+++ apps/OrderProcessor/src/main/java/com/tradingplatform/orderprocessor/routers/UsersRouter.java
@@ -1,13 +1,7 @@
-package com.tradingplatform.orderprocessor;
+package com.tradingplatform.orderprocessor.routers;
 
-import static com.tradingplatform.orderprocessor.database.DatabaseUtils.*;
-
-import com.mongodb.client.MongoCollection;
 import com.mongodb.client.model.Filters;
-import com.tradingplatform.orderprocessor.database.DatabaseUtils;
 import com.tradingplatform.orderprocessor.database.MongoClientConnection;
-import com.tradingplatform.orderprocessor.orders.Order;
-import com.tradingplatform.orderprocessor.orders.OrderType;
 import io.vertx.core.Vertx;
 import io.vertx.core.json.JsonObject;
 import io.vertx.ext.web.Router;
@@ -16,28 +10,22 @@ import java.util.Collections;
 import java.util.List;
 import org.bson.Document;
 
-public class OrderProcessorRouter {
+public class UsersRouter {
 
   private final Router router;
-  private final OrderProcessorService orderProcessorService;
+
+  UsersRouter(Vertx vertx) {
+    router = Router.router(vertx);
+    setupRoutes();
+  }
 
   public Router getRouter() {
     return router;
   }
 
-  OrderProcessorRouter(
-    Vertx vertx,
-    OrderProcessorService orderProcessorService
-  ) {
-    router = Router.router(vertx);
-    this.orderProcessorService = orderProcessorService;
-
-    initialise();
-  }
-
-  void initialise() {
+  private void setupRoutes() {
     router
-      .post("/create-user")
+      .post("/create")
       .handler(ctx -> {
         JsonObject body = ctx.getBodyAsJson();
         String userId = body.getString("userId");
@@ -63,7 +51,7 @@ public class OrderProcessorRouter {
       });
 
     router
-      .post("/update-user-balance")
+      .post("/update-balance")
       .handler(ctx -> {
         JsonObject body = ctx.getBodyAsJson();
         if (
@@ -85,7 +73,7 @@ public class OrderProcessorRouter {
       });
 
     router
-      .get("/user-active-positions")
+      .get("/active-positions")
       .handler(ctx -> {
         String userId = ctx.request().getParam("userId");
 
@@ -114,114 +102,7 @@ public class OrderProcessorRouter {
       });
 
     router
-      .post("/create-order")
-      .handler(ctx -> {
-        JsonObject body = ctx.getBodyAsJson();
-
-        MongoCollection<Document> activeOrdersCollection =
-          MongoClientConnection.getCollection("activeOrders");
-        MongoCollection<Document> orderHistoryCollection =
-          MongoClientConnection.getCollection("orderHistory");
-        MongoCollection<Document> usersCollection =
-          MongoClientConnection.getCollection("users");
-
-        String validationResult = passValidations(body, usersCollection);
-
-        if (!validationResult.equals("PASSED VALIDATIONS")) {
-          ctx
-            .response()
-            .setStatusCode(422)
-            .putHeader("Content-Type", "application/json")
-            .end(new JsonObject().put("Error", validationResult).encode());
-          return;
-        }
-
-        Order order = DatabaseUtils.createOrder(body);
-
-        insertOrderIntoDatabase(order, activeOrdersCollection);
-
-        ArrayList<Document> matchesFoundAsMongoDBDocs =
-          DatabaseUtils.processOrderAndParseMatchesFound(
-            order,
-            orderProcessorService
-          );
-
-        if (!matchesFoundAsMongoDBDocs.isEmpty()) {
-          updateDb(
-            matchesFoundAsMongoDBDocs,
-            activeOrdersCollection,
-            usersCollection,
-            orderHistoryCollection
-          );
-        }
-
-        ctx
-          .response()
-          .setStatusCode(201)
-          .putHeader("Content-Type", "application/json")
-          .end("Order created");
-      });
-
-    router
-      .post("/cancel-order")
-      .handler(ctx -> {
-        JsonObject body = ctx.getBodyAsJson();
-
-        String orderId = body.getString("orderId");
-        String userId = body.getString("userId");
-        String ticker = body.getString("ticker");
-        String type = body.getString("type");
-
-        var activeOrdersCollection = MongoClientConnection.getCollection(
-          "activeOrders"
-        );
-
-        var orderHistoryCollection = MongoClientConnection.getCollection(
-          "orderHistory"
-        );
-
-        var usersCollection = MongoClientConnection.getCollection("users");
-
-        JsonObject orderJson = JsonObject.mapFrom(
-          activeOrdersCollection.find(Filters.eq("orderId", orderId)).first()
-        );
-
-        Order order = DatabaseUtils.createOrder(orderJson);
-
-        orderProcessorService.cancelOrder(orderId, ticker, type);
-
-        activeOrdersCollection.deleteOne(Filters.eq("orderId", orderId));
-
-        if (order.getType() == OrderType.BUY) {
-          int amountToCreditBack = (int) order.getPrice() * order.getQuantity();
-
-          usersCollection.updateOne(
-            Filters.eq("userId", userId),
-            new Document("$inc", new Document("balance", amountToCreditBack))
-          );
-        }
-
-        if (previouslyPartiallyFilled(orderId)) {
-          orderHistoryCollection.updateOne(
-            Filters.eq("orderId", orderId),
-            new Document("$set", new Document("cancelled", true))
-          );
-        } else {
-          Document orderDoc = order.toDoc();
-          orderDoc.put("cancelled", true);
-
-          orderHistoryCollection.insertOne(orderDoc);
-        }
-
-        ctx
-          .response()
-          .setStatusCode(201)
-          .putHeader("Content-Type", "application/json")
-          .end("Order cancelled");
-      });
-
-    router
-      .get("/user-account")
+      .get("/account")
       .handler(ctx -> {
         String userId = ctx.request().getParam("userId");
         if (userId == null) {
@@ -254,7 +135,7 @@ public class OrderProcessorRouter {
       });
 
     router
-      .get("/user-trade-history")
+      .get("/trade-history")
       .handler(ctx -> {
         String userId = ctx.request().getParam("userId");
         if (userId == null) {
